+++ app/src/main/java/com/snowflowerthon/snowman/vote/VoteFragment.kt
@@ -1,0 +1,28 @@
+package com.snowflowerthon.snowman.vote
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.snowflowerthon.snowman.R
+import com.snowflowerthon.snowman.databinding.FragmentVoteBinding
+
+
+class VoteFragment : Fragment() {
+   lateinit var fragmentVoteBinding: FragmentVoteBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_vote, container, false)
+    }
+
+
+}
